refactor package controllers to web
